get rid of operatorkit
