--- flag/service/alertmanager/alertmanager.go
@@ -1,7 +1,0 @@
-package alertmanager
-
-type AlertManager struct {
-	Address        string
-	Authentication string
-	TenantId       string
-}
