+++ pkg/alertmanager/alertmanager.go
@@ -6,10 +6,18 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"time"
 
 	"github.com/giantswarm/microerror"
+	"github.com/pkg/errors"
 
-	"github.com/giantswarm/silence-operator/service/controller/key"
+	"github.com/giantswarm/silence-operator/api/v1alpha1"
+)
+
+const (
+	CreatedBy                = "silence-operator"
+	ValidUntilAnnotationName = "valid-until"
+	DateOnlyLayout           = "2006-01-02"
 )
 
 // TODO Get rid of microerrors and use errors instead.
@@ -30,7 +38,7 @@ type AlertManager struct {
 
 func New(config Config) (*AlertManager, error) {
 	if config.Address == "" {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Address must not be empty", config)
+		return nil, errors.Errorf("%T.Address must not be empty", config)
 	}
 
 	return &AlertManager{
@@ -102,7 +110,7 @@ func (am *AlertManager) DeleteSilenceByComment(comment string) error {
 	}
 
 	for _, s := range silences {
-		if s.Comment == comment && s.CreatedBy == key.CreatedBy {
+		if s.Comment == comment && s.CreatedBy == CreatedBy {
 			return am.DeleteSilenceByID(s.ID)
 		}
 	}
@@ -177,4 +185,36 @@ func (am *AlertManager) DeleteSilenceByID(id string) error {
 	}
 
 	return nil
+}
+
+func SilenceComment(silence *v1alpha1.Silence) string {
+	return fmt.Sprintf("%s-%s", CreatedBy, silence.Name)
+}
+
+// SilenceEndsAt gets the expiry date for a given silence.
+// The expiry date is retrieved from the annotation name configured by ValidUntilAnnotationName.
+// The expected format is defined by DateOnlyLayout.
+// It returns an invalidExpirationDateError in case the date format is invalid.
+func SilenceEndsAt(silence *v1alpha1.Silence) (time.Time, error) {
+	annotations := silence.GetAnnotations()
+
+	// Check if the annotation exist otherwise return a date 100 years in the future.
+	value, ok := annotations[ValidUntilAnnotationName]
+	if !ok {
+		return silence.GetCreationTimestamp().AddDate(100, 0, 0), nil
+	}
+
+	// Parse the date found in the annotation using RFC3339 (ISO8601) by default
+	expirationDate, err := time.Parse(time.RFC3339, value)
+	if err != nil {
+		// If it fails, we try to parse it using date only (old way)
+		expirationDate, err = time.Parse(DateOnlyLayout, value)
+		if err != nil {
+			return time.Time{}, microerror.Maskf(invalidExpirationDateError, "%s date %q does not match expected format %q: %v", ValidUntilAnnotationName, value, DateOnlyLayout, err)
+		}
+		// We shift the time to 8am UTC (9 CET or 10 CEST) to ensure silences do not expire at night.
+		expirationDate = time.Date(expirationDate.Year(), expirationDate.Month(), expirationDate.Day(), 8, 0, 0, 0, time.UTC)
+	}
+
+	return expirationDate, nil
 }
