+++ pkg/alertmanager/error.go
@@ -25,3 +25,19 @@ var notFoundError = &microerror.Error{
 func IsNotFound(err error) bool {
 	return microerror.Cause(err) == notFoundError
 }
+
+var wrongTypeError = &microerror.Error{
+	Kind: "wrongTypeError",
+}
+
+func IsWrongType(err error) bool {
+	return microerror.Cause(err) == wrongTypeError
+}
+
+var invalidExpirationDateError = &microerror.Error{
+	Kind: "invalidExpirationDateError",
+}
+
+func IsInvalidExpirationDate(err error) bool {
+	return microerror.Cause(err) == invalidExpirationDateError
+}
