--- server/endpoint/endpoint.go
@@ -1,56 +1,0 @@
-package endpoint
-
-import (
-	"github.com/giantswarm/microendpoint/endpoint/healthz"
-	"github.com/giantswarm/microendpoint/endpoint/version"
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-
-	"github.com/giantswarm/silence-operator/service"
-)
-
-type Config struct {
-	Logger  micrologger.Logger
-	Service *service.Service
-}
-
-type Endpoint struct {
-	Healthz *healthz.Endpoint
-	Version *version.Endpoint
-}
-
-func New(config Config) (*Endpoint, error) {
-	var err error
-
-	var healthzEndpoint *healthz.Endpoint
-	{
-		c := healthz.Config{
-			Logger: config.Logger,
-		}
-
-		healthzEndpoint, err = healthz.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var versionEndpoint *version.Endpoint
-	{
-		c := version.Config{
-			Logger:  config.Logger,
-			Service: config.Service.Version,
-		}
-
-		versionEndpoint, err = version.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	e := &Endpoint{
-		Healthz: healthzEndpoint,
-		Version: versionEndpoint,
-	}
-
-	return e, nil
-}
