--- server/server.go
@@ -1,97 +1,0 @@
-// Package server provides a server implementation to connect network transport
-// protocols and service business logic by defining server endpoints.
-package server
-
-import (
-	"context"
-	"net/http"
-	"sync"
-
-	"github.com/giantswarm/microerror"
-	microserver "github.com/giantswarm/microkit/server"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/viper"
-
-	"github.com/giantswarm/silence-operator/pkg/project"
-	"github.com/giantswarm/silence-operator/server/endpoint"
-	"github.com/giantswarm/silence-operator/service"
-)
-
-type Config struct {
-	Logger  micrologger.Logger
-	Service *service.Service
-
-	Viper *viper.Viper
-}
-
-func New(config Config) (microserver.Server, error) {
-	var err error
-
-	var endpointCollection *endpoint.Endpoint
-	{
-		c := endpoint.Config{
-			Logger:  config.Logger,
-			Service: config.Service,
-		}
-
-		endpointCollection, err = endpoint.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	s := &server{
-		logger: config.Logger,
-
-		bootOnce: sync.Once{},
-		config: microserver.Config{
-			Logger:      config.Logger,
-			ServiceName: project.Name(),
-			Viper:       config.Viper,
-
-			Endpoints: []microserver.Endpoint{
-				endpointCollection.Healthz,
-				endpointCollection.Version,
-			},
-			ErrorEncoder: encodeError,
-		},
-		shutdownOnce: sync.Once{},
-	}
-
-	return s, nil
-}
-
-type server struct {
-	logger micrologger.Logger
-
-	bootOnce     sync.Once
-	config       microserver.Config
-	shutdownOnce sync.Once
-}
-
-func (s *server) Boot() {
-	s.bootOnce.Do(func() {
-		// Here goes your custom boot logic for your server/endpoint if
-		// any.
-	})
-}
-
-func (s *server) Config() microserver.Config {
-	return s.config
-}
-
-func (s *server) Shutdown() {
-	s.shutdownOnce.Do(func() {
-		// Here goes your custom shutdown logic for your
-		// server/endpoint if any.
-	})
-}
-
-func encodeError(ctx context.Context, err error, w http.ResponseWriter) {
-	rErr := err.(microserver.ResponseError)
-	uErr := rErr.Underlying()
-
-	rErr.SetCode(microserver.CodeInternalError)
-	rErr.SetMessage(uErr.Error())
-	w.WriteHeader(http.StatusInternalServerError)
-}
